+++ django_machine_auth/__init__.py
@@ -1,6 +1,6 @@
 default_app_config = "django_machine_auth.apps.DjangoMachineAuthConfig"
 
-__version__ = "0.1.0"
+__version__ = "0.3.0"
 
 __all__ = [
     "__version__",