+++ django_machine_auth/serializers/permission.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from django_machine_auth.models import MachinePermission
+
+
+class MachinePermissionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = MachinePermission
+        fields = ["id", "module", "permission", "label"]