+++ django_machine_auth/views/__init__.py
@@ -1,2 +1,11 @@
-from .base_viewset import MachineAuthViewSet
 from .api_key_management_viewset import MachineAPIKeyManagementViewSet
+from .base_viewset import MachineAuthViewSet
+from .permission_viewset import MachinePermissionViewSet
+from .request_log_viewset import MachineAPIKeyRequestLogViewSet
+
+__all__ = [
+    "MachineAuthViewSet",
+    "MachineAPIKeyManagementViewSet",
+    "MachinePermissionViewSet",
+    "MachineAPIKeyRequestLogViewSet",
+]